Scrape from a more generic product list
There are some products which are not correctly listed. Take the Samsung
Galaxy Buds live [0] as an example, they are not listed under the sound
accessories despite being earphones. In order to workaround this issue,
we are going to process all accessories.

There's also a fix for the products that do not include the host on
href.

[0] https://loja.negocios.vodafone.pt/loja/acessorios/samsung/galaxy-buds-live-branco?src=ON