Update to match Vodafone API changes
The new API is much more friendly than before. It now returns a JSON
object with all the products. There is no pagination!
It also provides more information, but the little information the
scraper has been extracting has proven to be enough. Besides, we are
better protected in case there are breaking API changes in the future.

This is a nice update. Fewer lines of code and simplified logic while
keeping the same functionality.